requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
appnope==0.1.3
argcomplete==1.10.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.8.2
bleach==6.0.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
click-log==0.4.0
colorama==0.4.6
comm==0.1.3
compressed-rtf==1.0.6
contourpy==1.0.7
cycler==0.11.0
dataclasses==0.6
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
docutils==0.20.1
docx2txt==0.8
dotty-dict==1.3.1
ebcdic==1.1.1
executing==1.2.0
extract-msg==0.28.7
faiss-cpu==1.7.4
fastjsonschema==2.17.1
filelock==3.12.0
Flask==2.3.2
fonttools==4.39.4
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
gitdb==4.0.10
GitPython==3.1.31
gotrue==1.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.15.1
idna==2.10
IMAPClient==2.1.0
importlib-metadata==6.6.0
iniconfig==2.0.0
invoke==1.7.3
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keyring==23.13.1
kiwisolver==1.4.4
langchain==0.0.188
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
more-itertools==9.1.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.24.3
olefile==0.46
openai==0.27.7
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20191110
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkginfo==1.9.6
platformdirs==3.5.1
pluggy==1.0.0
postgrest==0.10.6
postgrest-py==0.4.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.8
Pygments==2.15.1
pyparsing==3.0.9
pypdf==3.9.0
pyrsistent==0.19.3
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==3.14.0
python-json-logger==2.0.7
python-pptx==0.6.21
python-semantic-release==7.33.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
readme-renderer==37.3
realtime==1.0.0
realtime-py==0.1.3
regex==2023.5.5
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
semver==2.13.0
Send2Trash==1.8.2
six==1.12.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.8.1
SQLAlchemy==2.0.15
stack-data==0.6.2
storage3==0.5.2
StrEnum==0.4.10
supabase==1.0.3
supabase-py==0.0.2
supafunc==0.2.2
tenacity==8.2.2
terminado==0.17.1
textract==1.6.5
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomlkit==0.11.8
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
twine==3.8.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
uri-template==1.2.0
urllib3==1.26.16
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
websockets==10.4
Werkzeug==2.3.4
widgetsnbextension==4.0.7
xlrd==1.2.0
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
